Add files via upload
This project is part of the final assignment for the Data Mining and Advanced Data Analysis course. The goal of this project is to build a predictive model for car prices based on a dataset of car listings. The project involves data preprocessing, feature engineering, model training, and feature importance analysis.